requirements.txt
Django==2.0.1
django-allauth==0.35.0
django-cors-headers==2.2.0d
djangorestframework==3.7.7
django-rest-auth==0.9.3
ipdb==0.10.3
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
whitenoise==3.3.1
psycopg2==2.7.3.2
gunicorn==19.6.0
dj-database-url==0.5.0
django-storages==1.6.5
boto3==1.6.5